requirements.txt
Django>=3.0,<4.0
djangorestframework ==3.13.1
psycopg2>=2.8
django-redis
